requirements.txt
telethon
gspread
google-auth